rag/additionnal-requirements.txt
chromadb
langchainhub
beautifulsoup4 
